changes the comments to it's own component and adds it to the topics pages
